MN Install Instructions for OSX Wallet Users
Detailed instructions on how to get your MN running if you are running the OSX version of the Trittium wallet. I will update the instructions with how to do from a fresh install as well as how to downgrade if already installed to the latest version of the daemon once the automated script I uploaded is approved and uploaded here to Github.